+++ sprockets/http/__init__.py
@@ -1,13 +1,16 @@
 import logging
 import logging.config
 import os
+import warnings
 
 
 version_info = (2, 0, 1)
 __version__ = '.'.join(str(v) for v in version_info)
 
+_unspecified = object()
 
-def run(create_application, settings=None, log_config=None):
+
+def run(create_application, settings=None, log_config=_unspecified):
     """
     Run a Tornado create_application.
 
@@ -61,8 +64,16 @@ def run(create_application, settings=None, log_config=None):
     debug_mode = bool(app_settings.get('debug',
                                        int(os.environ.get('DEBUG', 0)) != 0))
     app_settings['debug'] = debug_mode
-    logging.config.dictConfig(_get_logging_config(debug_mode)
-                              if log_config is None else log_config)
+    if log_config is _unspecified:
+        warnings.warn(
+            'calling sprockets.http.run without logging configuration is '
+            'deprecated and will fail in the future', DeprecationWarning)
+        logging.config.dictConfig(_get_logging_config(debug_mode))
+        logging.warning(
+            'calling sprockets.http.run without logging configuration is '
+            'deprecated and will fail in the future')
+    else:
+        logging.config.dictConfig(log_config)
 
     port_number = int(app_settings.pop('port', os.environ.get('PORT', 8000)))
     num_procs = int(app_settings.pop('number_of_procs', '0'))